contracts: formatting

+++ contracts/contracts/tests/Gas.sol
@@ -7,15 +7,11 @@ import {Sapphire} from "../Sapphire.sol";
 contract GasTests {
     bytes32 tmp;
 
-    function testConstantTime(uint useGas, uint128 padGasAmount)
-        external
-    {
-        if( useGas == 1 )
-        {
+    function testConstantTime(uint256 useGas, uint128 padGasAmount) external {
+        if (useGas == 1) {
             bytes32 x;
 
-            for( uint i = 0; i < 100; i++ )
-            {
+            for (uint256 i = 0; i < 100; i++) {
                 x = keccak256(abi.encodePacked(x, tmp));
             }
